feat: add agent bulilder & page builder & group chat support upload files
