+++ src/features/ChatInput/Desktop/index.tsx
@@ -6,11 +6,7 @@ import { createStyles } from 'antd-style';
 import { ReactNode, memo, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import DragUploadZone from '@/components/DragUploadZone';
 import { useChatInputStore } from '@/features/ChatInput/store';
-import { useModelSupportVision } from '@/hooks/useModelSupportVision';
-import { useAgentStore } from '@/store/agent';
-import { agentByIdSelectors } from '@/store/agent/selectors';
 import { useChatStore } from '@/store/chat';
 import { chatSelectors } from '@/store/chat/selectors';
 import { fileChatSelectors, useFileStore } from '@/store/file';
@@ -18,7 +14,6 @@ import { useGlobalStore } from '@/store/global';
 import { systemStatusSelectors } from '@/store/global/selectors';
 
 import ActionBar, { type ActionToolbarProps } from '../ActionBar';
-import { useAgentId } from '../hooks/useAgentId';
 import InputEditor from '../InputEditor';
 import SendArea from '../SendArea';
 import TypoBar from '../TypoBar';
@@ -81,25 +76,6 @@ const DesktopChatInput = memo<DesktopChatInputProps>(
 
     const chatKey = useChatStore(chatSelectors.currentChatKey);
 
-    // Get agent info for vision support check
-    const agentId = useAgentId();
-    const model = useAgentStore((s) => agentByIdSelectors.getAgentModelById(agentId)(s));
-    const provider = useAgentStore((s) => agentByIdSelectors.getAgentModelProviderById(agentId)(s));
-    const canUploadImage = useModelSupportVision(model, provider);
-    const uploadFiles = useFileStore((s) => s.uploadChatFiles);
-
-    const handleUploadFiles = async (files: File[]) => {
-      // Filter out image files if the model does not support vision
-      const filteredFiles = files.filter((file) => {
-        if (canUploadImage) return true;
-        return !file.type.startsWith('image');
-      });
-
-      if (filteredFiles.length > 0) {
-        uploadFiles(filteredFiles);
-      }
-    };
-
     useEffect(() => {
       if (editor) editor.focus();
     }, [chatKey, editor]);
@@ -109,52 +85,50 @@ const DesktopChatInput = memo<DesktopChatInputProps>(
     const contextContainerNode = shouldShowContextContainer && <ContextContainer />;
 
     return (
-      <DragUploadZone onUploadFiles={handleUploadFiles}>
-        <Flexbox
-          className={cx(styles.container, expand && styles.fullscreen)}
-          gap={8}
-          paddingBlock={expand ? 0 : showFootnote ? '0 12px' : '0 16px'}
+      <Flexbox
+        className={cx(styles.container, expand && styles.fullscreen)}
+        gap={8}
+        paddingBlock={expand ? 0 : showFootnote ? '0 12px' : '0 16px'}
+      >
+        <ChatInput
+          defaultHeight={chatInputHeight || 32}
+          footer={
+            <ChatInputActionBar
+              left={<ActionBar dropdownPlacement={dropdownPlacement} />}
+              right={<SendArea />}
+              style={{
+                paddingRight: 8,
+              }}
+            />
+          }
+          fullscreen={expand}
+          header={
+            <Flexbox gap={0}>
+              {extenHeaderContent}
+              {showTypoBar && <TypoBar />}
+              {contextContainerNode}
+            </Flexbox>
+          }
+          maxHeight={320}
+          minHeight={36}
+          onSizeChange={(height) => {
+            updateSystemStatus({ chatInputHeight: height });
+          }}
+          resize={true}
+          slashMenuRef={slashMenuRef}
+          {...inputContainerProps}
+          className={cx(expand && styles.inputFullscreen, inputContainerProps?.className)}
         >
-          <ChatInput
-            defaultHeight={chatInputHeight || 32}
-            footer={
-              <ChatInputActionBar
-                left={<ActionBar dropdownPlacement={dropdownPlacement} />}
-                right={<SendArea />}
-                style={{
-                  paddingRight: 8,
-                }}
-              />
-            }
-            fullscreen={expand}
-            header={
-              <Flexbox gap={0}>
-                {extenHeaderContent}
-                {showTypoBar && <TypoBar />}
-                {contextContainerNode}
-              </Flexbox>
-            }
-            maxHeight={320}
-            minHeight={36}
-            onSizeChange={(height) => {
-              updateSystemStatus({ chatInputHeight: height });
-            }}
-            resize={true}
-            slashMenuRef={slashMenuRef}
-            {...inputContainerProps}
-            className={cx(expand && styles.inputFullscreen, inputContainerProps?.className)}
-          >
-            <InputEditor />
-          </ChatInput>
-          {showFootnote && !expand && (
-            <Center style={{ pointerEvents: 'none', zIndex: 100 }}>
-              <Text className={styles.footnote} type={'secondary'}>
-                {t('input.disclaimer')}
-              </Text>
-            </Center>
-          )}
-        </Flexbox>
-      </DragUploadZone>
+          <InputEditor />
+        </ChatInput>
+        {showFootnote && !expand && (
+          <Center style={{ pointerEvents: 'none', zIndex: 100 }}>
+            <Text className={styles.footnote} type={'secondary'}>
+              {t('input.disclaimer')}
+            </Text>
+          </Center>
+        )}
+      </Flexbox>
     );
   },
 );
